spinq_backend/requirements.txt
Django==2.1.2
django-versatileimagefield==1.10
Pillow==5.3.0
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
